+++ hi_modules/modulators/mods/ControlModulator.cpp
@@ -158,7 +158,7 @@ void ControlModulator::setInternalAttribute (int parameter_index, float newValue
 		{
 		defaultValue = newValue;
 
-		handleHiseEvent(HiseEvent(HiseEvent::Type::Controller, controllerNumber, (int)defaultValue, 1));
+		handleHiseEvent(HiseEvent(HiseEvent::Type::Controller, (uint8)controllerNumber, (uint8)defaultValue, 1));
 
 		break;
